improves long plugin system message to the LLM
