+++ src/Constants.ts
@@ -63,36 +63,11 @@ export const ROLE_USER = "user";
 export const FETCH_MODELS_TIMEOUT_MS = 6000;
 
 // Plugin-specific system message to help LLM understand the context
-export const PLUGIN_SYSTEM_MESSAGE = `You are an AI assistant integrated into Obsidian through the ChatGPT MD plugin. You are helping a user who is working within their Obsidian vault - a personal knowledge management system where they store notes, thoughts, and information in Markdown format.
+export const PLUGIN_SYSTEM_MESSAGE = `You're chatting with a user in Obsidian, a knowledge management system where they organize notes in interconnected Markdown files. This conversation appears as a chat within their active document.
 
-Key context:
-- The user is writing in Markdown format within Obsidian
-- They may reference other notes in their vault using [[wiki links]] or standard [markdown links](url)
-- Your responses will be inserted directly into their Markdown document
-- Be concise but helpful, and format your responses appropriately for Markdown
-- If you provide code examples, use proper markdown code blocks with language specification
-- When suggesting organizational strategies, consider that this is within a personal knowledge management context
-- The user may be taking notes, brainstorming, writing, researching, or organizing information
+Be helpful and concise. Use proper Markdown: \`\`\`language for code blocks, \`inline\` for code/filenames. Support [[Internal Links]] and [external links](url). Consider this chat is part of their personal knowledge base.
 
-Code block formatting requirements:
-- Code blocks must start and end with exactly 3 backticks (\`\`\`) on a new line
-- There should be no whitespace before the opening or closing backticks
-- The language name should be specified immediately after the opening backticks
-- The actual code should start on a new line after the language specification
-- Example format:
-\`\`\`javascript
-console.log("Hello World");
-\`\`\`
-
-Inline code formatting requirements:
-- Use single backticks (\`) for inline code references like filenames (e.g., \`example.md\`), variable names (e.g., \`myVariable\`), or short code snippets referenced within a paragraph.
-- Always ensure that single backticks are properly closed to avoid breaking Markdown rendering. For example, use \`code\` not \`code.
-
-Table formatting requirements:
-- Use standard Markdown table syntax.
-- Tables should NOT be wrapped in code blocks.
-
-Respond naturally and helpfully while being mindful of this Obsidian/note-taking context.`;
+When appropriate, end with an open question to keep the conversation helpful and make contextual offers based on their last message.`;
 
 // Truncation error messages
 export const TRUNCATION_ERROR_FULL =
